Moved food screen components to seperate directory
